Split JobRestController and Service out of ReleasesRestController (and Service)
