+++ butler/src/main/java/rocks/metaldetector/butler/facade/ReleaseService.java
@@ -1,6 +1,5 @@
 package rocks.metaldetector.butler.facade;
 
-import rocks.metaldetector.butler.facade.dto.ImportJobResultDto;
 import rocks.metaldetector.butler.facade.dto.ReleaseDto;
 import rocks.metaldetector.support.Page;
 import rocks.metaldetector.support.PageRequest;
@@ -13,12 +12,6 @@ public interface ReleaseService {
   Page<ReleaseDto> findReleases(Iterable<String> artists, TimeRange timeRange, String query, PageRequest pageRequest);
 
   List<ReleaseDto> findAllReleases(Iterable<String> artists, TimeRange timeRange);
-
-  void createImportJob();
-
-  void createRetryCoverDownloadJob();
-
-  List<ImportJobResultDto> queryImportJobResults();
 
   void updateReleaseState(long releaseId, String state);
 }
